+++ x/emissions/types/validations.go
@@ -1103,6 +1103,17 @@ func (msg *RegisterRequest) Validate() error {
 	return nil
 }
 
+// validate that a transfer actor ownership request follows the expected format
+func (msg *TransferActorOwnershipRequest) Validate() error {
+	if _, err := sdk.AccAddressFromBech32(msg.Sender); err != nil {
+		return errors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid sender address (%s)", err)
+	}
+	if _, err := sdk.AccAddressFromBech32(msg.NewOwner); err != nil {
+		return errors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid new owner address (%s)", err)
+	}
+	return nil
+}
+
 // validate that a remove registration request follows the expected format
 func (msg *RemoveRegistrationRequest) Validate() error {
 	_, err := sdk.AccAddressFromBech32(msg.Sender)